cratedb: Refactor schema.Evolution
