+++ debezium-server-cratedb/src/main/java/io/debezium/server/cratedb/schema/Schema.java
@@ -14,17 +14,17 @@ import java.util.stream.Collectors;
 
 public record Schema() {
 
+    public sealed interface I permits Array, Bit, Coli, Dict, Primitive {
+    }
+
     public static I of() {
         return Dict.of();
     }
-
     public enum Primitive implements I {
         BIGINT,
         BOOLEAN,
-        TEXT
-    }
-
-    public sealed interface I permits Array, Bit, Coli, Dict, Primitive {
+        TEXT,
+        TIMETZ
     }
 
     public record Array(I innerType) implements I {
